Simulated Data
The underlying data generating process incorporates reciprocity as a driving force behind the effort provision in the experiment's second stage. Many other variables are simulated randomly or using some shortcuts to approximate predicted behavior.

The database.xml file shares the format of the actual experiment's output files. It can be processed with the 01_Read_XML_Data.R script. The simulatedExperimentData.csv file is the document that is generated by that code.