Codebook
This codebook lists all of the variables from the experiment (for data the raw as well as processed data). The codebook has two sections. In the first section, variable name, label, and some additional information are given. It does not provide any summary statistics or information about the data types, however. Where applicable, the second part of the codebook provides text of the survey questions.