Create 06_SubSample.R
This file repeats parts of the manipulation, the whole Analysis and Visualization of the other Scripts but uses a different sample. In particular, the data only considers the observations where the agent's sensitivity to reciprocity (YAgent) is above median.

It is rather messy as it overwrites the objects previously stored in your R console. 
The corresponding outputs (figures and tables) are stored separately with the prefix "RS" for Reciprocity Sample.