Adjusted IT manipulation to new design
Under the old design, there were two likelihoods which we interpret as monitoring intensity. Under the new design these likelihoods are either 100% or 0%. In the experiment file we defined the variable as a binary variable with values 1 (for a performance based payoff) or 2 (for a random payoff). I therefore adjusted the R script accordingly.